warnings fixed
git-svn-id: https://svn.vis.uni-stuttgart.de/projects/megamol/plugins/protein/trunk@333 5e8c2511-817d-0410-85a1-fd227d6480d1

+++ GridNeighbourFinder.h
@@ -53,7 +53,7 @@ namespace protein {
 				/* resize bounding-box and grid structure */
 				vislib::math::Dimension<T, 3> dim = boundingBox.GetSize();
 				for(int i = 0; i < 3; i++)
-					this->gridResolution[i] =  floor(dim[i] / (2*searchDistance) + 1.0);
+					this->gridResolution[i] =  (unsigned int)floor(dim[i] / (2*searchDistance) + 1.0);
 				this->gridSize = this->gridResolution[0] * this->gridResolution[1] * this->gridResolution[2];
 				this->elementBBox = boundingBox;
 
@@ -61,12 +61,12 @@ namespace protein {
 				if (elementGrid)
 					delete [] elementGrid;
 				this->elementGrid = new vislib::Array<const T *>[gridSize];
-				for (int i = 0; i < gridSize; i++) {
+				for (unsigned int i = 0; i < gridSize; i++) {
 					elementGrid[i].Resize( 100);
 					elementGrid[i].SetCapacityIncrement( 100);
 				}
 			} else {
-				for (int i = 0; i < gridSize; i++ )
+				for (unsigned int i = 0; i < gridSize; i++ )
 					elementGrid[i].Clear();
 			}
 
@@ -79,7 +79,7 @@ namespace protein {
 				this->cellSize[i] = (T)bBoxDimension[i] / this->gridResolution[i]; //(T)1.0) / gridResolutionFactors[i];
 			}
 			// sort the element positions into the grid ...
-			for(int i = 0; i < this->elementCount; i++) {
+			for(unsigned int i = 0; i < this->elementCount; i++) {
 				// we can skip this assert becaus another ASSERT in 'insertPointIntoGrid'
 				//ASSERT( elementBBox.Contains(vislib::math::ShallowPoint<T,3>(const_cast<T*>(&elementPositions[i*3]))) );
 				if (!filter || filter[i] != -1)
@@ -97,13 +97,13 @@ namespace protein {
 
 			// calculate range in the grid ...
 			int min[3], max[3];
-			for(int i = 0; i < 3; i++) {
-				min[i] = floor((relPos[i]-distance)*gridResolutionFactors[i]);
+			for(unsigned int i = 0; i < 3; i++) {
+				min[i] = (int)floor((relPos[i]-distance)*gridResolutionFactors[i]);
 				if (min[i] < 0)
 					min[i] = 0;
-				max[i] = ceil((relPos[i]+distance)*gridResolutionFactors[i]);
-				if (max[i] >= gridResolution[i])
-					max[i] = gridResolution[i]-1;
+				max[i] = (int)ceil((relPos[i]+distance)*gridResolutionFactors[i]);
+				if (max[i] >= (int)gridResolution[i])
+					max[i] = (int)gridResolution[i]-1;
 			}
 
 			// loop over all cells inside the sphere (point, distance)
@@ -128,9 +128,9 @@ namespace protein {
 	private:
 		VISLIB_FORCEINLINE void insertPointIntoGrid(const T *point) {
 			//Point relPos = sub(point, elementOrigin);
-			unsigned int indexX = /*relPos.X()*/(point[0] - elementOrigin[0]) * gridResolutionFactors[0]; // floor()?
-			unsigned int indexY = /*relPos.Y()*/(point[1] - elementOrigin[1]) * gridResolutionFactors[1];
-			unsigned int indexZ = /*relPos.Z()*/(point[2] - elementOrigin[2]) * gridResolutionFactors[2];
+			unsigned int indexX = (unsigned int)(/*relPos.X()*/(point[0] - elementOrigin[0]) * gridResolutionFactors[0]); // floor()?
+			unsigned int indexY = (unsigned int)(/*relPos.Y()*/(point[1] - elementOrigin[1]) * gridResolutionFactors[1]);
+			unsigned int indexZ = (unsigned int)(/*relPos.Z()*/(point[2] - elementOrigin[2]) * gridResolutionFactors[2]);
 			ASSERT(indexX < gridResolution[0] && indexY < gridResolution[1] && indexZ < gridResolution[2]);
 			vislib::Array<const T *>& cell = elementGrid[cellIndex(indexX, indexY, indexZ)];
 			cell.Add(point);
@@ -139,7 +139,7 @@ namespace protein {
 		VISLIB_FORCEINLINE void findNeighboursInCell(const vislib::Array<const T *>& cell, const T* point, T distance, vislib::Array<unsigned int>& resIdx) const {
 			for(int i = 0; i < cell.Count(); i++)
 				if ( dist(cell[i],point) <= distance )
-					resIdx.Add((cell[i]-elementPositions)/3); // store atom index
+					resIdx.Add((unsigned int)((cell[i]-elementPositions)/3)); // store atom index
 		}
 
 		inline unsigned int cellIndex(unsigned int x, unsigned int y, unsigned int z) const {
